Everything is here except for testing

+++ StregSystemCore/IStregsystem.cs
@@ -8,7 +8,7 @@ namespace StregsystemCore
 {
     public interface IStregsystem {
         IEnumerable<BaseProduct> ActiveProducts { get; }
-        InsertCashTransaction AddCreditsToAccount(User user, int amount);
+        InsertCashTransaction AddCreditsToAccount(User user, decimal amount);
         BuyTransaction BuyProduct(User user, BaseProduct product);
         BaseProduct GetProductByID(int id);
         IEnumerable<Transaction> GetTransactions(User user, int count);
